requirements.txt
Flask
gunicorn
pillow
numpy
tensorflow
keras
werkzeug
google-cloud-storage